[#1614] Refactor the code
Group together all the duplicated code
